Move Spinner from lab to core and add documentation
