+++ .changeset/sharp-terms-doubt.md
@@ -1,0 +1,6 @@
+---
+"@salt-ds/core": minor
+"@salt-ds/lab": minor
+---
+
+Move Spinner from lab to core
