+++ packages/ag-grid-theme/stories/examples/InfiniteScroll.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from "react";
 import { AgGridReact, AgGridReactProps } from "ag-grid-react";
-import { StackLayout } from "@salt-ds/core";
-import { Spinner } from "@salt-ds/lab";
+import { Spinner, StackLayout } from "@salt-ds/core";
 import dataGridExampleData from "../dependencies/dataGridExampleData";
 import dataGridInfiniteScrollExampleColumns from "../dependencies/dataGridInfiniteScrollExampleColumns";
 import { useAgGridHelpers } from "../dependencies/useAgGridHelpers";
