+++ packages/ag-grid-theme/stories/examples/LoadingOverlay.tsx
@@ -1,7 +1,6 @@
 import { CSSProperties } from "react";
 import { AgGridReact, AgGridReactProps } from "ag-grid-react";
-import { Card, StackLayout } from "@salt-ds/core";
-import { Spinner } from "@salt-ds/lab";
+import { Card, Spinner, StackLayout } from "@salt-ds/core";
 import dataGridExampleData from "../dependencies/dataGridExampleData";
 import dataGridExampleColumns from "../dependencies/dataGridExampleColumns";
 import { useAgGridHelpers } from "../dependencies/useAgGridHelpers";
