+++ packages/core/stories/spinner/spinner.doc.stories.mdx
@@ -1,0 +1,61 @@
+import { ArgsTable, Meta, Story, Canvas } from "@storybook/addon-docs";
+import { Spinner } from "@salt-ds/core";
+import {
+  CharacteristicUsage,
+  CSSClassTable,
+} from "css-variable-docgen-components";
+import HelpAndSupport from "docs/blocks/help-and-support.doc.mdx";
+
+<Meta
+  title="Documentation/Core/Spinner"
+  component={Spinner}
+  parameters={{
+    viewMode: "docs",
+  }}
+/>
+
+# Spinner
+
+A Spinner is a visual representation of a process that is taking an indeterminate time to complete. This can be triggered by automated system processes, user action, or to indicate that content is loading. This reassures the user that the system is still working, especially when the operation is taking longer than 300ms.
+
+### Best practice
+
+Improve accessibility by customizing the aria-label to provide additional context about _what_ is loading, e.g. `aria-label="loading settings panel"`.
+
+## Default
+
+Spinners have two sizes, `default` and `large`.
+
+<Canvas>
+  <Story id="core-spinner--default" />
+</Canvas>
+
+## Large
+
+Set the `size` prop to `"large"` to use the large Spinner.
+
+<Canvas>
+  <Story id="core-spinner--large" />
+</Canvas>
+
+## Configuring Spinner
+
+### API
+
+```
+import { Button } from "@salt-ds/spinner";
+```
+
+### Props
+
+<ArgsTable of={Spinner} />
+
+### Classes
+
+<CSSClassTable of={Spinner} />
+
+### Characteristics used
+
+<CharacteristicUsage of={Spinner} />
+
+<HelpAndSupport />
