+++ packages/lab/src/index.ts
@@ -61,7 +61,6 @@ export * from "./scrim";
 export * from "./search-input";
 export * from "./skip-link";
 export * from "./slider";
-export * from "./spinner";
 export * from "./stepper-input";
 export * from "./switch";
 export * from "./tabs";
